% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{clusterCCs}
\alias{clusterCCs}
\title{Simple clustering scheme linking all detections within a search radius}
\usage{
clusterCCs(coords, r)
}
\arguments{
\item{coords}{A matrix containing coordinates of the detections.}

\item{r}{The search radius used to contruct the graph.}
}
\value{
Vector containing the cluster indices for each detection, a cluster index of zero refers to noise
}
\description{
Forms a graph by linking all detections within a specified search radius. Clusters defined as connected components of this graph.
}
