% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{clusterDBSCAN}
\alias{clusterDBSCAN}
\title{Density-based spatial clustering of applications with noise (DBSCAN)}
\usage{
clusterDBSCAN(coords, r, minPoints)
}
\arguments{
\item{coords}{A matrix containing coordinates of the detections.}

\item{r}{The search radius for density estimation}

\item{minPoints}{Minimum number of detections that should be within the search radius for a detection to be defined as a "core" point.}
}
\value{
Vector containing the cluster indices for each detection, a cluster index of zero refers to noise
}
\description{
As orginally described in Ester, Martin, et al. "Density-based spatial clustering of applications with noise." Int. Conf. Knowledge Discovery and Data Mining. Vol. 240. 1996.
This function simply calls the dbscan function from the library dbscan and is included for convenience. Border detections are included in clusters.
}
