% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{clusterKDE}
\alias{clusterKDE}
\title{2D Kernel Density Estimate (KDE) based clustering}
\usage{
clusterKDE(coords, std, threshold, gridWidth = 5)
}
\arguments{
\item{coords}{A matrix containing coordinates of the detections.}

\item{std}{The standard deviation of the Guassian kernels. Either a fixed value, a vector of length 2, a vector of length equal to the number of detecions or a matrix with the same dimension as coords.}

\item{threshold}{Threshold for density esimate. Between 0 and 1.}

\item{gridWidth}{Defines the resolution of the evulation grid. Default value of 5.}
}
\value{
Vector containing the cluster indices for each detection, a cluster index of zero refers to noise
}
\description{
Uses a Guassian kernel to estimate detection detensity on a grid. Estimate is thresholded and the connected components define the cluster. Threshold is scaled between the minimum and maxium of the density estimate.
}
