% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{clusterPersitenceRipley}
\alias{clusterPersitenceRipley}
\title{Persistence clustering using the Ripley L function to define the detection density}
\usage{
clusterPersitenceRipley(coords, r, threshold, ROIArea)
}
\arguments{
\item{coords}{A matrix containing coordinates of the detections.}

\item{r}{The search radius for density estimation and to contruct linking graph}

\item{threshold}{Persistence threshold for L-r}

\item{ROIArea}{Area of the field of view. Needed to calcualte the Ripley K function.}
}
\value{
Vector containing the cluster indices for each detection, a cluster index of zero refers to noise
}
\description{
Persistence clustering using the Ripley L function to define the detection density
}
