% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{clusterRipley}
\alias{clusterRipley}
\title{Ripley's K function based clustering}
\usage{
clusterRipley(coords, r, threshold, ROIArea)
}
\arguments{
\item{coords}{A matrix containing coordinates of the detections.}

\item{r}{The search radius for calcualte of the Ripley K and L functions.}

\item{threshold}{Threshold for L-r.}

\item{ROIArea}{Area of the field of view. Needed to calcualte the Ripley K function.}
}
\value{
Vector containing the cluster indices for each detection, a cluster index of zero refers to noise
}
\description{
First the number of detections within a fixed search radius, r, is counted. This is used to claculate the Ripley K function which is normalised to get the Ripley L function.
Detections with L-r less than a specified threshold are removed. Clusters are defined by the connected components of the graph formed by linking all filtered detections within r.
This is approximatelly as described in the following paper but without Baysian parameter selection: \cr \cr
Rubin-Delanchy, Patrick, et al. "Bayesian cluster identification in single-molecule localization microscopy data." Nature methods 12.11 (2015): 1072.
}
