% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{clusterStats}
\alias{clusterStats}
\title{Calculate per cluster statsitics}
\usage{
clusterStats(coords, clusterIndices, areaMethod = "ch", scale = 0)
}
\arguments{
\item{coords}{A matrix containing coordinates of the detections.}

\item{clusterIndices}{The cluster index for each detection. Noise detections should have cluster index zero.}
}
\value{
A data frame containing statistics for each cluster
}
\description{
Calculates the number of detections, area, volume density and centre of mass for each cluster. Returns statistics for each cluster and also mean values.
}
