% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{clusterTomato}
\alias{clusterTomato}
\title{Topological Model Analysis Tool (ToMATo) clustering}
\usage{
clusterTomato(coords, r, threshold)
}
\arguments{
\item{coords}{A matrix containing coordinates of the detections.}

\item{r}{The search radius for density estimation and to contruct linking graph}

\item{threshold}{Persistence threshold for merging of density modes}
}
\value{
Vector containing the cluster indices for each detection, a cluster index of zero refers to noise
}
\description{
As orginally described in Chazal, Frederic, et al. "Persistence-based clustering in riemannian manifolds." Journal of the ACM (JACM) 60.6 (2013): 41.
}
