% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{clusterVoronoi}
\alias{clusterVoronoi}
\title{2D Voronoi tesselation based clustering}
\usage{
clusterVoronoi(coords, threshold, densityChoice)
}
\arguments{
\item{coords}{A matrix containing coordinates of the detections.}

\item{threshold}{Threshold for density.}

\item{densityChoice}{Choice of density calculation method. 0: inverse tile area, 1: inverse mean first rank tile area}
}
\value{
Vector containing the cluster indices for each detection, a cluster index of zero refers to noise
}
\description{
A Voronoi tesselation is contructed from the detection coordinates. Detection density is defined as the inverse of tile area, or mean first rank tile area.
Detections with density below a specified threshold are removed from the dual Delaunay triangulation and the connected components of this graph are used to define clusters. 
Implementation is for 2D data only. \cr \cr
Levet, Florian, et al. "SR-Tesseler: a method to segment and quantify localization-based super-resolution microscopy data." Nature methods 12.11 (2015): 1065. \cr
Andronov, Leonid, et al. "ClusterViSu, a method for clustering of protein complexes by Voronoi tessellation in super-resolution microscopy." Scientific reports 6 (2016): 24084.
}
