% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{filtClustArea}
\alias{filtClustArea}
\title{Filter clusters based on area/volume}
\usage{
filtClustArea(coords, clusterIndices, minArea, maxArea)
}
\arguments{
\item{coords}{A matrix containing coordinates of the detections.}

\item{clusterIndices}{The cluster index for each detection. Noise detections should have cluster index zero.}

\item{minArea}{Minimum cluster area/volume.}

\item{maxArea}{Maximum cluster area/volume.}
}
\value{
Vector containing filtered cluster indices
}
\description{
In 3D clusters are filtered by volume instead of area.
Area/volume defined by convex hull. 
Detections in clusters outside specified range are asigned an index of zero.
}
