% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{filtClustDetections}
\alias{filtClustDetections}
\title{Filter clusters based on number of detecions in a cluster}
\usage{
filtClustDetections(clusterIndices, minDetections, maxDetections)
}
\arguments{
\item{clusterIndices}{The cluster index for each detection. Noise detections should have cluster index zero.}

\item{minDetections}{The minimum number of detections each cluster should contain.}

\item{maxDetections}{The maximum number of detections each cluster should contain.}
}
\value{
Vector containing filtered cluster indices
}
\description{
Detections in clusters outside specified range are asigned an index of zero
}
