% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistentHomologyUtilities.R
\name{findNumFeatures}
\alias{findNumFeatures}
\title{Utility function to return the number of features above a specified threshold from a persistence diagram.}
\usage{
findNumFeatures(diag, dimension, threshold)
}
\arguments{
\item{diag}{3 Column matrix containing the perisitence diagram.  Rows correspond to individual features. Columns correspond to feature dimension, birth scale and death scale.}

\item{dimension}{The dimension of features to search for (0: connected components, 1: holes, 2: voids).}

\item{threshold}{Persitence threshold for features.}
}
\value{
The number of features of the specified dimension above the specified threshold.
}
\description{
Utility function to return the number of features above a specified threshold from a persistence diagram.
}
