% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistentHomologyUtilities.R
\name{findNumFeaturesConsensus}
\alias{findNumFeaturesConsensus}
\title{Givin a list of diagrams calculated the number of features above a specified threshold and calculates the modal value and frequency}
\usage{
findNumFeaturesConsensus(diags, dimensions, thresholds)
}
\arguments{
\item{diag}{3 Column matrix containing the perisitence diagram.  Rows correspond to individual features. Columns correspond to feature dimension, birth scale and death scale.}

\item{dimension}{The dimension of features to search for (0: connected components, 1: holes, 2: voids).}

\item{threshold}{Persitence threshold for features.}
}
\value{
List containing the modal number of features (numFeatures) and the fraction of diagrams which find the modal number (agreement)
}
\description{
Givin a list of diagrams calculated the number of features above a specified threshold and calculates the modal value and frequency
}
