% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{plotClusterScatter}
\alias{plotClusterScatter}
\title{A simple 2D plotting function using ggplot2 to display clustering results.}
\usage{
plotClusterScatter(coords, clusterIndices)
}
\arguments{
\item{coords}{A matrix containing x and y coordinates of the detections.}

\item{clusterIndices}{The cluster index for each detection. Noise detections should have cluster index zero.}
}
\description{
Distinct cluster colours are sampled from a selection produced by randomColoR
}
