% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{plotTomatoDiagram}
\alias{plotTomatoDiagram}
\title{Simple plotting function for ToMATo diagrams}
\usage{
plotTomatoDiagram(diagram, threshold)
}
\arguments{
\item{diagram}{Matix containing the bith and death densities for all modes in the density estimate}

\item{threshold}{Persistence threshold for merging of density modes}
}
\description{
Simple plotting function for ToMATo diagrams
}
