% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistentHomologyUtilities.R
\name{sampledDiagrams}
\alias{sampledDiagrams}
\title{Sample detection coordinates (with) replacement a set number of times. For each set of sample coordinates compute the persitence diagram.}
\usage{
sampledDiagrams(coords, maxDimension, maxScale, numBootstraps = 50,
  samplingRate = 1, weightingParameter, weightingFactor = 2.302585,
  numberInGroup)
}
\arguments{
\item{coords}{A matrix containing coordinates of the detections.}

\item{maxDimension}{The maximum dimenstion of topological features to calculate (0: CCs, 1: CCs/holes, 2: CCs/holes/voids).}

\item{maxScale}{Maximum scale over which to calculate the filtration}

\item{numBootstraps}{number of repeats for sub-sampling process. Default 20.}

\item{samplingRate}{A number between 0 and 1 specify the rate at which to sample the coordinates. Default 1.}

\item{weightingParameter}{(Optional) Parameter used to weight the sampling. For example the localization precision.}

\item{weightingFactor}{Posive number specifying the weighting factor for exponential decay function. Default log(0.1).}

\item{numberInGroup}{(Optional) Number of detections grouped into each specified coordinate.}
}
\value{
List containing the persistent diagrams from the sampled data.
}
\description{
Sample detection coordinates (with) replacement a set number of times. For each set of sample coordinates compute the persitence diagram.
}
