% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationUtilities.R
\name{simulateSMLM}
\alias{simulateSMLM}
\title{Simulates SMLM imaging process to generate synthetic single molecule localization microscopy datasets.}
\usage{
simulateSMLM(labelCoords, blinkGeomProb, precisionMeanLog, precisionSdLog,
  detectionRate, falseDetectionRate, fieldLimits, precisionMeanLogNoise,
  precisionSdLogNoise, labelIndex)
}
\arguments{
\item{labelCoords}{A matrix containing x and y (and z in 3D) coordinates of the underlying label distribution.}

\item{blinkGeomProb}{Probablilty of transistion to a dark state. Used to generate a geometric distrubtion.}

\item{precisionMeanLog}{Mean for lognormal distrubtion used to determine molecule localisation precision.}

\item{precisionSdLog}{Standard deviation for lognormal distrubtion used to determine molecule localisation precision.}

\item{detectionRate}{Percentage of blinking events which are detected.}

\item{falseDetectionRate}{Percentage of noise detections to add.}

\item{fieldLimits}{A matrix containing the field size limits.}

\item{precisionMeanLogNoise}{(Optional) Mean for lognormal distrubtion used to determine localisation precision of noisy detections.}

\item{precisionSdLogNoise}{(Optional) Standard for lognormal distrubtion used to determine localisation precision of noisy detections.}

\item{labelIndex}{(Optional) Vector linking labels to molecules}
}
\value{
Data frame containing the detection list (coordinates for all detections), and also the molecule list (linking detections to the orginating molecules)
}
\description{
Simulates SMLM imaging process to generate synthetic single molecule localization microscopy datasets.
}
