% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationUtilities.R
\name{simulateSTORM}
\alias{simulateSTORM}
\title{Simulates dSTORM imaging process to generate synthetic single molecule localization microscopy datasets}
\usage{
simulateSTORM(moleculeCoords, averageLabelsPerMol, blinkGeomProb,
  precisionMeanLog, precisionSdLog, detectionRate, falseDetectionRate,
  fieldLimit)
}
\arguments{
\item{moleculeCoords}{A matrix containing x and y (and z in 3D) coordinates of the underlying molecule distribution.}

\item{averageLabelsPerMol}{Average number of fluophores attatched to each molecule.}

\item{blinkGeomProb}{Probablilty of transistion to a dark state. Used to generate a geometric distrubtion.}

\item{precisionMeanLog}{Mean for lognormal distrubtion used to determine molecule localisation precision.}

\item{precisionSdLog}{Standard deviation for lognormal distrubtion used to determine molecule localisation precision.}

\item{detectionRate}{Percentage of blinking events which are detected.}

\item{falseDetectionRate}{Percentage of noise detections to add.}

\item{fieldLimit}{A matrix containing the field size limits.}
}
\value{
Data frame containing the detection list (coordinates for all detections), and also the molecule list (linking detections to the orginating molecules)
}
\description{
Simulates dSTORM imaging process to generate synthetic single molecule localization microscopy datasets
}
