% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtilities.R
\name{tomatoDiagram}
\alias{tomatoDiagram}
\title{Produces a ToMATo persistence diagram}
\usage{
tomatoDiagram(coords, r)
}
\arguments{
\item{coords}{A matrix containing coordinates of the detections.}

\item{r}{The search radius for density estiamtion and to contruct linking graph}
}
\value{
Matix containing the bith and death densities for all modes in the density estimate
}
\description{
As orginally described in Chazal, Frederic, et al. "Persistence-based clustering in riemannian manifolds." Journal of the ACM (JACM) 60.6 (2013): 41.
Records the bith and death densities for all density modes
Calculated by running the ToMATo clustering with very high threshold
}
